StateAnchored를 제거 하고 StateHidden 추가
