+++ hover/src/main/java/io/mattcarroll/hover/HoverViewStateType.java
@@ -5,5 +5,5 @@ public enum HoverViewStateType {
     COLLAPSED,
     PREVIEWED,
     EXPANDED,
-    ANCHORED,
+    HIDDEN,
 }
